Restructured PrismService without servlets
